Change double to decimal

+++ GospoRol.Domain/Models/Warehouse.cs
@@ -6,9 +6,9 @@ namespace GospoRol.Domain.Models
     public class Warehouse : BaseEntity
     {
         public string Name { get; set; }
-        public double Acreage { get; set; }                 //areał // ładowność
-        public double AcreageFree { get; set; }             //Wolny areał
-        public double AcreageOccupied { get; set; }         //zajęty Areał
+        public decimal Acreage { get; set; }                 //areał // ładowność
+        public decimal AcreageFree { get; set; }             //Wolny areał
+        public decimal AcreageOccupied { get; set; }         //zajęty Areał
         public virtual ICollection<Product> Products { get; set; }
         //
         // maszyny + płody rolne +
